stateをuint32_tに変更

+++ src/global.hpp
@@ -14,7 +14,7 @@ extern int32_t user_mem[1024];
 extern word dict_mem[256];
 
 extern void* stack_base;
-extern uint8_t state;
+extern uint32_t state;
 extern void* here;
 
 #endif //GLOBAL_HPP
